+++ web/modules/custom/hoeringsportal_citizen_proposal/modules/hoeringsportal_citizen_proposal_fixtures/src/Fixture/CitizenProposalFixture.php
@@ -9,6 +9,7 @@ use Drupal\Core\Datetime\DrupalDateTime;
 use Drupal\hoeringsportal_base_fixtures\Fixture\MediaFixture;
 use Drupal\hoeringsportal_base_fixtures\Fixture\ParagraphFixture;
 use Drupal\hoeringsportal_base_fixtures\Helper\Helper as BaseFixtureHelper;
+use Drupal\hoeringsportal_citizen_proposal\Form\ProposalFormBase;
 use Drupal\hoeringsportal_citizen_proposal\Helper\Helper;
 use Drupal\hoeringsportal_citizen_proposal\Helper\MailHelper;
 use Drupal\hoeringsportal_citizen_proposal_archiving\Helper\Helper as ArchiveHelper;
@@ -51,20 +52,30 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
       'field_author_uuid' => '1111',
       'field_author_name' => 'Anders And',
       'field_author_email' => 'anders.and@itkdev.dk',
-      'field_vote_start' => DrupalDateTime::createFromFormat('U', strtotime('tomorrow'))->format('Y-m-d\TH:i:s'),
-      'field_vote_end' => DrupalDateTime::createFromFormat('U', strtotime('tomorrow +3 months'))->format('Y-m-d\TH:i:s'),
       'field_content_state' => 'upcoming',
       'field_proposal' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlShort.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
       'field_remarks' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlShort.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
     ]);
     $entity->save();
+
+    // field_vote_start and field_vote_end are set on first publish, so we set
+    // it after saving once.
+    $entity->field_vote_start->setValue(DrupalDateTime::createFromFormat('U', strtotime('tomorrow'))->format('Y-m-d\TH:i:s'));
+    $entity->field_vote_end->setValue(DrupalDateTime::createFromFormat('U', strtotime('tomorrow +3 months'))->format('Y-m-d\TH:i:s'));
+    $entity->save();
+
     $this->addReference('node:citizen_proposal:Proposal1', $entity);
+
+    // Add some support.
+    for ($i = 0; $i < 87; $i++) {
+      $this->helper->saveSupport(uniqid('', TRUE), $entity, ['user_name' => self::class]);
+    }
 
     $entity = Node::create([
       'type' => 'citizen_proposal',
@@ -73,20 +84,29 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
       'field_author_uuid' => '2222',
       'field_author_name' => 'Fedtmule',
       'field_author_email' => 'fedtmule@itkdev.dk',
-      'field_vote_start' => DrupalDateTime::createFromFormat('U', strtotime('yesterday -3 months'))->format('Y-m-d\TH:i:s'),
-      'field_vote_end' => DrupalDateTime::createFromFormat('U', strtotime('yesterday'))->format('Y-m-d\TH:i:s'),
       'field_content_state' => 'finished',
       'field_proposal' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtml1.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
       'field_remarks' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlLong.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
     ]);
     $entity->save();
+
+    $entity->field_vote_start->setValue(DrupalDateTime::createFromFormat('U', strtotime('yesterday -3 months'))->format('Y-m-d\TH:i:s'));
+    $entity->field_vote_end->setValue(DrupalDateTime::createFromFormat('U', strtotime('yesterday'))->format('Y-m-d\TH:i:s'));
+    $entity->save();
+
     $this->addReference('node:citizen_proposal:Proposal2', $entity);
+
+    // Add some support.
+    for ($i = 0; $i < 3; $i++) {
+      $this->helper->saveSupport(uniqid('', TRUE), $entity, ['user_name' => self::class]);
+    }
+    $this->helper->saveSupport(uniqid('', TRUE), $entity, ['user_name' => 'Ø']);
 
     $entity = Node::create([
       'type' => 'citizen_proposal',
@@ -95,16 +115,14 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
       'field_author_uuid' => '3333',
       'field_author_name' => 'Hexia De Trick',
       'field_author_email' => 'givmiglykkemønten@itkdev.dk',
-      'field_vote_start' => DrupalDateTime::createFromFormat('U', strtotime('-1 month'))->format('Y-m-d\TH:i:s'),
-      'field_vote_end' => DrupalDateTime::createFromFormat('U', strtotime('+2 months'))->format('Y-m-d\TH:i:s'),
       'field_content_state' => 'active',
       'field_proposal' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlLong.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
       'field_remarks' => [
         'value' => $this->baseFixtureHelper->getText('filteredHtmlLong.html'),
-        'format' => 'filtered_html',
+        'format' => ProposalFormBase::CONTENT_TEXT_FORMAT,
       ],
     ]);
     $entity->save();
@@ -113,7 +131,17 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
     // Set admin values.
     $data = Yaml::parseFile(__DIR__ . '/CitizenProposalFixture/citizen_proposal_admin_form_values.yaml');
     if (isset($data['citizen_proposal_admin_form_values'])) {
-      $this->helper->setAdminValues($data['citizen_proposal_admin_form_values']);
+      $values = $data['citizen_proposal_admin_form_values'];
+
+      /** @var \Drupal\node\Entity\NodeInterface $node */
+      $node = $this->getReference('node:static_page:thanks');
+      $values['approve_goto_url'] = $node->toUrl(options: ['alias' => TRUE])->toString();
+
+      /** @var \Drupal\node\Entity\NodeInterface $node */
+      $node = $this->getReference('node:landing_page:Proposals');
+      $values['cancel_goto_url'] = $node->toUrl(options: ['alias' => TRUE])->toString();
+
+      $this->helper->setAdminValues($values);
     }
   }
 
@@ -124,6 +152,8 @@ class CitizenProposalFixture extends AbstractFixture implements DependentFixture
     return [
       MediaFixture::class,
       ParagraphFixture::class,
+      CitizenProposalLandingPageFixture::class,
+      CitizenProposalStaticPageFixture::class,
     ];
   }