+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Form/ProposalFormAdd.php
@@ -10,6 +10,8 @@ use Drupal\node\Entity\Node;
  * Form for adding proposal.
  */
 final class ProposalFormAdd extends ProposalFormBase {
+  public const SURVEY_KEY = 'create_proposal_survey';
+
   public const ADD_FORM_TITLE_MAXLENGTH = 80;
   public const ADD_FORM_PROPOSAL_MAXLENGTH = 2000;
   public const ADD_FORM_REMARKS_MAXLENGTH = 10000;
@@ -58,6 +60,11 @@ final class ProposalFormAdd extends ProposalFormBase {
       '#description_display' => 'before',
     ];
 
+    if ($this->isAuthenticatedAsEditor()) {
+      $form['name']['#required'] = TRUE;
+      unset($form['name']['#attributes']['readonly']);
+    }
+
     $form['phone'] = [
       '#type' => 'textfield',
       '#required' => TRUE,
@@ -85,7 +92,14 @@ final class ProposalFormAdd extends ProposalFormBase {
         ->t('Display email'),
       '#default_value' => $defaltValues['email_display'] ?? TRUE,
       '#description' => $this->getAdminFormStateValue('email_display_help'),
-      '#description_display' => 'before',
+    ];
+
+    $form['allow_email'] = [
+      '#type' => 'checkbox',
+      '#title' => $this
+        ->t('Allow email'),
+      '#default_value' => $defaltValues['allow_email'] ?? FALSE,
+      '#description' => $this->getAdminFormStateValue('allow_email_help'),
     ];
 
     $form['proposal_intro_container'] = [
@@ -115,14 +129,16 @@ final class ProposalFormAdd extends ProposalFormBase {
     ];
 
     $form['proposal'] = [
-      '#type' => 'textarea',
+      '#type' => 'text_format',
+      '#format' => self::CONTENT_TEXT_FORMAT,
+      '#allowed_formats' => [self::CONTENT_TEXT_FORMAT],
       '#required' => TRUE,
       '#rows' => 15,
       '#title' => $this
         ->t('Proposal'),
       '#description' => $this->getAdminFormStateValue('proposal_help'),
       '#description_display' => 'before',
-      '#default_value' => $defaltValues['proposal'],
+      '#default_value' => $defaltValues['proposal'] ?? '',
       '#maxlength_js' => TRUE,
       '#attributes' => [
         'data-maxlength' => $this->getMaxLength('characters_proposal'),
@@ -131,20 +147,24 @@ final class ProposalFormAdd extends ProposalFormBase {
     ];
 
     $form['remarks'] = [
-      '#type' => 'textarea',
+      '#type' => 'text_format',
+      '#format' => self::CONTENT_TEXT_FORMAT,
+      '#allowed_formats' => [self::CONTENT_TEXT_FORMAT],
       '#required' => TRUE,
       '#rows' => 15,
       '#title' => $this
         ->t('Remarks'),
       '#description' => $this->getAdminFormStateValue('remarks_help'),
       '#description_display' => 'before',
-      '#default_value' => $defaltValues['remarks'],
+      '#default_value' => $defaltValues['remarks'] ?? '',
       '#maxlength_js' => TRUE,
       '#attributes' => [
         'data-maxlength' => $this->getMaxLength('characters_remarks'),
         'maxlength_js_label' => $this->t('@remaining characters left.'),
       ],
     ];
+
+    $this->buildSurveyForm($form);
 
     $form['consent'] = [
       '#type' => 'checkbox',
@@ -164,6 +184,34 @@ final class ProposalFormAdd extends ProposalFormBase {
       '#button_type' => 'primary',
     ];
 
+    $form['#after_build'][] = $this->afterBuildForm(...);
+
+    return $form;
+  }
+
+  /**
+   * Form after build handler.
+   *
+   * @param array $form
+   *   The form.
+   * @param \Drupal\Core\Form\FormStateInterface $formState
+   *   The form state.
+   *
+   * @return array
+   *   The form.
+   */
+  private function afterBuildForm(array $form, FormStateInterface $formState) {
+    // Hide text format info
+    // (lifted from _allowed_formats_remove_textarea_help()).
+    foreach ($form as $key => &$element) {
+      if (isset($element['format'])) {
+        unset(
+          $element['format']['help'],
+          $element['format']['guidelines']
+        );
+      }
+    }
+
     return $form;
   }
 
@@ -175,11 +223,20 @@ final class ProposalFormAdd extends ProposalFormBase {
       $form_state->setErrorByName('title', $this->t('Too many characters used.'));
     }
 
-    if (strlen($form_state->getValue('proposal')) > $this->getMaxLength('characters_proposal')) {
+    $getTextValue = function (string $key) use ($form_state) {
+      $value = $form_state->getValue($key);
+      if (isset($value['value'])) {
+        $value = $value['value'];
+      }
+
+      return strip_tags((string) $value);
+    };
+
+    if (strlen($getTextValue('proposal')) > $this->getMaxLength('characters_proposal')) {
       $form_state->setErrorByName('proposal', $this->t('Too many characters used.'));
     }
 
-    if (strlen($form_state->getValue('remarks')) > $this->getMaxLength('characters_remarks')) {
+    if (strlen($getTextValue('remarks')) > $this->getMaxLength('characters_remarks')) {
       $form_state->setErrorByName('remarks', $this->t('Too many characters used.'));
     }
   }
@@ -196,18 +253,15 @@ final class ProposalFormAdd extends ProposalFormBase {
       'field_author_phone' => $formState->getValue('phone'),
       'field_author_email' => $formState->getValue('email'),
       'field_author_email_display' => $formState->getValue('email_display'),
-      'field_proposal' => [
-        'value' => $formState->getValue('proposal'),
-        'format' => 'filtered_html',
-      ],
-      'field_remarks' => [
-        'value' => $formState->getValue('remarks'),
-        'format' => 'filtered_html',
-      ],
+      'field_proposal' => $formState->getValue('proposal'),
+      'field_remarks' => $formState->getValue('remarks'),
+      'field_author_allow_email' => $formState->getValue('allow_email'),
     ]);
     $this->helper->setDraftProposal($entity);
     $formState
       ->setRedirect('hoeringsportal_citizen_proposal.citizen_proposal.proposal_approve');
+
+    $this->setSurveyResponse($formState);
   }
 
   /**