+++ web/modules/custom/hoeringsportal_citizen_proposal/src/Form/ProposalFormBase.php
@@ -10,8 +10,11 @@ use Drupal\Core\StringTranslation\TranslatableMarkup;
 use Drupal\Core\Url;
 use Drupal\hoeringsportal_citizen_proposal\Exception\RuntimeException;
 use Drupal\hoeringsportal_citizen_proposal\Helper\Helper;
+use Drupal\hoeringsportal_citizen_proposal\Helper\WebformHelper;
 use Drupal\hoeringsportal_openid_connect\Controller\OpenIDConnectController;
 use Drupal\hoeringsportal_openid_connect\Helper as AuthenticationHelper;
+use Drupal\node\NodeInterface;
+use Drupal\webform\WebformInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 
@@ -19,12 +22,16 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
  * Base form for adding proposal.
  */
 abstract class ProposalFormBase extends FormBase {
+  public const SURVEY_KEY = NULL;
+
+  public const CONTENT_TEXT_FORMAT = 'citizen_proposal_content';
 
   /**
    * Constructor for the proposal add form.
    */
   final public function __construct(
     readonly protected Helper $helper,
+    readonly protected WebformHelper $webformHelper,
     readonly private AuthenticationHelper $authenticationHelper,
     readonly private ImmutableConfig $config
   ) {
@@ -36,8 +43,9 @@ abstract class ProposalFormBase extends FormBase {
   public static function create(ContainerInterface $container) {
     return new static(
       $container->get(Helper::class),
+      $container->get(WebformHelper::class),
       $container->get(AuthenticationHelper::class),
-      $container->get('config.factory')->get('hoeringsportal_citizen_proposal.settings')
+      $container->get('hoeringsportal_citizen_proposal.config.settings')
     );
   }
 
@@ -137,6 +145,7 @@ abstract class ProposalFormBase extends FormBase {
       'title' => $entity?->title->value ?? NULL,
       'proposal' => $entity?->field_proposal->value ?? '',
       'remarks' => $entity?->field_remarks->value ?? '',
+      'allow_email' => $entity?->field_author_allow_email->value ?? FALSE,
     ];
   }
 
@@ -145,6 +154,18 @@ abstract class ProposalFormBase extends FormBase {
    */
   protected function getAdminFormStateValue(string|array $key, string $default = NULL): mixed {
     return $this->helper->getAdminValue($key, $default);
+  }
+
+  /**
+   * Get admin form value as a URL.
+   */
+  protected function getAdminFormStateValueUrl(string|array $key, string $default = NULL, Url $defaultUrl = NULL): Url {
+    try {
+      return Url::fromUserInput($this->helper->getAdminValue($key, $default) ?? '');
+    }
+    catch (\Exception) {
+      return $defaultUrl ?? Url::fromRoute('<front>');
+    }
   }
 
   /**
@@ -183,6 +204,27 @@ abstract class ProposalFormBase extends FormBase {
   }
 
   /**
+   * De-authenticate (is that a real word?) user.
+   */
+  protected function deAuthenticateUser(Url $url = NULL): Url {
+    if (NULL === $url) {
+      $url = Url::fromRoute('<current>');
+    }
+
+    if (!$this->isAuthenticatedAsCitizen()) {
+      return $url;
+    }
+
+    $this->authenticationHelper->removeUserData();
+    return Url::fromRoute(
+      'hoeringsportal_openid_connect.openid_connect_end_session',
+      [
+        OpenIDConnectController::QUERY_STRING_DESTINATION => $url->toString(TRUE)->getGeneratedUrl(),
+      ],
+    );
+  }
+
+  /**
    * Get user UUID.
    *
    * @return string
@@ -203,6 +245,93 @@ abstract class ProposalFormBase extends FormBase {
 
     // Compute a GDPR safe and (hopefully) unique user identifier.
     return sha1($userId);
+  }
+
+  /**
+   * Build survey form.
+   *
+   * @param array $form
+   *   The form.
+   */
+  protected function buildSurveyForm(array &$form) {
+    $webform = $this->loadSurvey();
+    if (NULL === $webform) {
+      return;
+    }
+
+    $form['survey'] = [
+      '#type' => 'container',
+      '#attributes' => [
+        'class' => ['survey', 'citizen-proposal-survey'],
+      ],
+      '#tree' => TRUE,
+    ];
+
+    try {
+      $description = $this->getAdminFormStateValue([
+        static::SURVEY_KEY,
+        'description',
+      ]);
+      if (isset($description['value'])) {
+        // We use a numeric index (implicit 0) here to prevent webform fields
+        // accidentally overwriting the description element.
+        $form['survey'][] = [
+          '#type' => 'processed_text',
+          '#text' => $description['value'],
+          '#format' => $description['format'] ?? 'filtered_html',
+        ];
+
+        $this->webformHelper->renderWebformElements($webform, $form['survey']);
+      }
+    }
+    catch (\Exception $exception) {
+      throw $exception;
+    }
+  }
+
+  /**
+   * Load survey webform.
+   *
+   * @return \Drupal\webform\WebformInterface|null
+   *   The webform if any.
+   */
+  protected function loadSurvey(): ?WebformInterface {
+    if (empty(static::SURVEY_KEY)) {
+      return NULL;
+    }
+
+    return $this->webformHelper->loadWebform((string) $this->getAdminFormStateValue([
+      static::SURVEY_KEY,
+      'webform',
+    ]));
+  }
+
+  /**
+   * Set survey response.
+   */
+  protected function setSurveyResponse(FormStateInterface $formState) {
+    try {
+      if ($webform = $this->loadSurvey()) {
+        $surveyData = (array) $formState->getValue('survey');
+        $this->webformHelper->setSurveyResponse($webform, $surveyData);
+      }
+    }
+    catch (\Exception) {
+    }
+  }
+
+  /**
+   * Save survey response previously set.
+   */
+  protected function saveSurveyResponse(NodeInterface $node) {
+    try {
+      if ($webform = $this->loadSurvey()) {
+        $this->webformHelper->saveSurveyResponse($webform, $node);
+      }
+    }
+    catch (\Exception) {
+    }
+
   }
 
 }