+++ CHANGELOG.md
@@ -6,6 +6,59 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+* [PR-367](https://github.com/itk-dev/hoeringsportal/pull/367)
+  Added and used Editor Advanced link
+
+## [3.5.0] - 2023-08-30
+
+* [PR-363](https://github.com/itk-dev/hoeringsportal/pull/363)
+  Fixed issue with storing names containing non-ascii characters
+* [PR-364](https://github.com/itk-dev/hoeringsportal/pull/364)
+  Fixed sorting of citizen proposals
+* [PR-365](https://github.com/itk-dev/hoeringsportal/pull/365)
+  Fixed use of private temp store
+
+## [3.4.0] - 2023-08-29
+
+* [PR-361](https://github.com/itk-dev/hoeringsportal/pull/361)
+  Added cancel proposal redirect URL
+* [PR-360](https://github.com/itk-dev/hoeringsportal/pull/360)
+  Cleaned up display of checkboxes
+* [PR-359](https://github.com/itk-dev/hoeringsportal/pull/359)
+  Generalized handling of form surveys. Added survey to support form
+* [PR-358](https://github.com/itk-dev/hoeringsportal/pull/358)
+  Fixed display of description and counter
+
+## [3.3.0] - 2023-08-25
+
+* [PR-355](https://github.com/itk-dev/hoeringsportal/pull/355)
+  Added citizen access check.
+* [PR-356](https://github.com/itk-dev/hoeringsportal/pull/356)
+  Miscellaneous fixes
+
+## [3.2.0] - 2023-08-23
+
+* [PR-353](https://github.com/itk-dev/hoeringsportal/pull/353)
+  Added “Allow email” checkboxes
+* [PR-351](https://github.com/itk-dev/hoeringsportal/pull/351)
+  Added and used citizen proposal text format
+* [PR-350](https://github.com/itk-dev/hoeringsportal/pull/350)
+  Added superfluous support proposal button
+* [PR-347](https://github.com/itk-dev/hoeringsportal/pull/347)
+  Added citizen proposal survey
+* [PR-342](https://github.com/itk-dev/hoeringsportal/pull/342)
+  Added email and storage consent checkbox on citizen proposal support form.
+  Added supporters view.
+* [PR-346](https://github.com/itk-dev/hoeringsportal/pull/346)
+  Made filter and sort labels visually hidden
+* [PR-344](https://github.com/itk-dev/hoeringsportal/pull/344)
+  Hid changed date on static page
+* [PR-345](https://github.com/itk-dev/hoeringsportal/pull/345)
+  Added support redirect URL
+* [PR-341](https://github.com/itk-dev/hoeringsportal/pull/341)
+  Removed superfluous support proposal button
+* [PR-340](https://github.com/itk-dev/hoeringsportal/pull/340)
+  Fixed proposal progress bar display
 * [PR-337](https://github.com/itk-dev/hoeringsportal/pull/337)
   Updated drupal core 9.5.10 and contrib modules
 * [PR-332](https://github.com/itk-dev/hoeringsportal/pull/332)
@@ -27,8 +80,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   * toolbar_visibility
   * twig_tweak
   * viewsreference
+* [PR-336](https://github.com/itk-dev/hoeringsportal/pull/336)
+  De-authenticated citizens after creating or supporting proposals
 
-## [3.1.0] - 2023-08-04 - Citizen proposal
+## [3.1.0] - 2023-08-04
 
 * [PR-331](https://github.com/itk-dev/hoeringsportal/pull/331)
   Bumped `itk-dev/openid-connect` version
@@ -69,7 +124,10 @@ Updated drupal core 8.6.16
 
 Initial release
 
-[Unreleased]: https://github.com/itk-dev/hoeringsportal/compare/3.1.0...HEAD
+[Unreleased]: https://github.com/itk-dev/hoeringsportal/compare/3.4.0...HEAD
+[3.4.0]: https://github.com/itk-dev/hoeringsportal/compare/3.3.0...3.4.0
+[3.3.0]: https://github.com/itk-dev/hoeringsportal/compare/3.2.0...3.3.0
+[3.2.0]: https://github.com/itk-dev/hoeringsportal/compare/3.1.0...3.2.0
 [3.1.0]: https://github.com/itk-dev/hoeringsportal/compare/3.0.2...3.1.0
 [3.0.2]: https://github.com/itk-dev/hoeringsportal/compare/3.0.1...3.0.2
 [3.0.1]: https://github.com/itk-dev/hoeringsportal/compare/3.0.0...3.0.1